Add functions OnCollisionWithPlatForm , Render , GetBoundingBox , SetState The CMushRoom::OnCollisionWithPlatForm(LPCOLLISIONEVENT e) method is called to handle collisions with the background (CPlatform). In this method:
First, we check if the collision object is CPlatform or not by converting the object type (dynamic_cast) from e->obj to CPlatform.
Then we check if the background is a blocking object (IsBlocking()). If true, no further processing is required.
If the collision occurs from the bottom (e->ny < 0), we reposition along the y-axis of the mushroom (SetY(platform->GetY() - MUSHROOM_BBOX_HEIGHT)) to place the mushroom on top of the background.
The CMushRoom::Render() method is called to draw mushrooms on the screen.
Check the model value of the mushroom to determine if the mushroom is red (MUSHROOM_RED) or green (MUSHROOM_GREEN) and draw the corresponding animation on the screen (animations->Get(ID_ANI_MUSHROOM_RED)->Render(x, y) or animations-> Get(ID_ANI_MUSHROOM_GREEN)->Render(x, y)).
The other function dont need to explain